Function_learning/requirements.txt
fastapi[all]
google-generativeai==0.2.2